+++ web/src/app/index.html
@@ -9,7 +9,7 @@
       height: 100%;
     }
     .intercom-messenger-frame {
-      width: 300px !important;
+      width: 320px !important;
     }
   </style>
 </head>
